+++ projects/server/config/config.json
@@ -1,7 +1,7 @@
 {
     "development": {
         "username": "root",
-        "password": "password",
+        "password": "mushroomss",
         "database": "shopedia",
         "host": "127.0.0.1",
         "dialect": "mysql",
