+++ projects/server/routes/adminRoute.js
@@ -1,0 +1,8 @@
+const adminController = require("../controllers/adminController")
+const express = require("express")
+
+const router = express.Router()
+
+router.post("/login", adminController.adminLogin)
+
+module.exports = router
