+++ projects/server/src/index.js
@@ -3,19 +3,18 @@ const express = require("express")
 const cors = require("cors")
 const { join } = require("path")
 const db = require("../models")
-const authRoute = require("../routes/authRoute")
+const adminRoute = require("../routes/adminRoute")
 
 const PORT = process.env.PORT || 8000
 const app = express()
 app.use(
-    cors(
-        //     {
-        //     origin: [
-        //         process.env.WHITELISTED_DOMAIN &&
-        //         process.env.WHITELISTED_DOMAIN.split(","),
-        //     ],
-        // }
-    )
+    cors()
+    //     {
+    //     origin: [
+    //         process.env.WHITELISTED_DOMAIN &&
+    //         process.env.WHITELISTED_DOMAIN.split(","),
+    //     ],
+    // }
 )
 
 app.use(express.json())
@@ -24,8 +23,7 @@ app.use(express.json())
 
 // ===========================
 // NOTE : Add your routes here
-
-app.use("/auth", authRoute)
+app.use("/admin", adminRoute)
 
 app.get("/api", (req, res) => {
     res.send(`Hello, this is my API`)
